Se corrige tipo de dato de precio a Double, se agregan endpoints de categorias

+++ crud_Marcelo_Iglesia/src/main/java/com/T_Tech_CRUD_M_iglesia/crud_Marcelo_Iglesia/services/CategoriaServiceImpl.java
@@ -45,6 +45,12 @@ public class CategoriaServiceImpl implements CategoriaService {
 
     @Override
     public void eliminarCategoria(Long id) {
+        Categoria categoria = categoriaRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Categoría no encontrada"));
+
+        if (categoria.getArticulos() != null && !categoria.getArticulos().isEmpty()) {
+            throw new RuntimeException("No se puede eliminar la categoría porque tiene artículos asociados");
+        }
         categoriaRepository.deleteById(id);
 
     }
